Moved to LML; new ViewSystem based on simple includes, reflection, and LML.

+++ core/src/main/java/com/draniksoft/ome/utils/respone/ResponseCode.java
@@ -2,10 +2,10 @@ package com.draniksoft.ome.utils.respone;
 
 public class ResponseCode {
 
-    public static final int SUCCESSFUL = 11;
-    public static final int FAILED = 12;
+    public static final short SUCCESSFUL = 11;
+    public static final short FAILED = 12;
 
-    public static final int PAUSED = 13;
+    public static final short PAUSED = 13;
 
     public static final int STEP_ONE = 31;
     public static final int STEP_TWO = 32;
